+++ lib/flutter_flip_flap.dart
@@ -4,5 +4,5 @@ export 'models/flip_flap_item.dart';
 export 'split_flap_theme.dart';
 export 'widgets/flap/flap_text_unit.dart';
 export 'widgets/flap/unit_tile.dart';
-export 'widgets/flip_flap_display.dart';
 export 'widgets/flip/flip_text_unit.dart';
+export 'widgets/flip_flap_display.dart';