+++ lib/widgets/flap/flap_widget_unit.dart
@@ -3,28 +3,22 @@ import 'package:flutter_flip_flap/widgets/core/back_out_curves.dart';
 import 'package:flutter_flip_flap/widgets/core/flap_animator.dart';
 import 'package:flutter_flip_flap/widgets/core/flap_controller_mixin.dart';
 import 'package:flutter_flip_flap/widgets/core/jitter_duration_mixin.dart';
+import 'package:flutter_flip_flap/widgets/core/unit_base.dart';
 
-class FlapWidgetUnit extends StatefulWidget {
+class FlapWidgetUnit extends FlapUnitBase {
   const FlapWidgetUnit({
     super.key,
     required this.child,
-    required this.unitConstraints,
-    this.unitDecoration,
-    this.duration = const Duration(milliseconds: 200),
-    this.durationJitterMs = 50,
-    this.textStyle,
-    this.enableBounce = true,
-    this.bounceOvershoot = 2.8,
+    required super.unitConstraints,
+    super.unitDecoration,
+    super.duration,
+    super.durationJitterMs,
+    super.textStyle,
+    super.enableBounce,
+    super.bounceOvershoot,
   });
 
-  final BoxConstraints unitConstraints;
-  final Decoration? unitDecoration;
   final Widget child;
-  final Duration duration;
-  final int durationJitterMs;
-  final TextStyle? textStyle;
-  final bool enableBounce;
-  final double bounceOvershoot;
 
   @override
   State<FlapWidgetUnit> createState() => _FlapWidgetUnitState();
@@ -32,7 +26,6 @@ class FlapWidgetUnit extends StatefulWidget {
 
 class _FlapWidgetUnitState extends State<FlapWidgetUnit>
     with TickerProviderStateMixin, JitterDurationMixin, FlapControllerMixin {
-
   late Widget _currentChild;
   late Widget _nextChild;
 
@@ -79,10 +72,7 @@ class _FlapWidgetUnitState extends State<FlapWidgetUnit>
     }
   }
 
-  Duration get _effectiveDuration => effectiveDuration(
-    base: widget.duration,
-    jitterMs: widget.durationJitterMs,
-  );
+  Duration get _effectiveDuration => effectiveDuration(base: widget.duration, jitterMs: widget.durationJitterMs);
 
   @override
   Duration get flapDuration => _effectiveDuration;
@@ -94,11 +84,7 @@ class _FlapWidgetUnitState extends State<FlapWidgetUnit>
       decoration: widget.unitDecoration,
       child: _currentChild,
     ),
-    nextFace: _UnitFace(
-      constraints: widget.unitConstraints,
-      decoration: widget.unitDecoration,
-      child: _nextChild,
-    ),
+    nextFace: _UnitFace(constraints: widget.unitConstraints, decoration: widget.unitDecoration, child: _nextChild),
     animation: flapAnimation,
     secondStage: flapSecondStage,
   );